Added cookies on to locals, and expose helpers
