+++ src/routes/session/index.ts
@@ -1,3 +1,4 @@
+// @ts-nocheck
 import type { RequestHandler } from '@sveltejs/kit';
 
 export const get: RequestHandler = ({ locals }) => {
